pixel-additional: Check DONT_COPY_BT_CERT before copying BT cert
* Some devices can't use Google's prebuilt Bluetooth APEX

* For those device that using AOSP Bluetooth APEX,
  don't copy BT certs to prevent sepolicy breakages

+++ vendorsetup.sh
@@ -13,5 +13,8 @@ repo sync -c --force-sync --optimized-fetch --no-tags --no-clone-bundle --prune 
 
 rm -Rf vendor/aosp/rro_overlays/GooglePermissionControllerOverlay
 
-rm -Rf vendor/partner_modules/build/certificates/bluetooth.x509.pem
-cp vendor/pixel-additional/products/apex/certificates/bluetooth.x509.pem vendor/partner_modules/build/certificates/
+if [ ! "$DONT_COPY_BT_CERT" = true ];
+then
+    rm -Rf vendor/partner_modules/build/certificates/bluetooth.x509.pem
+    cp vendor/pixel-additional/products/apex/certificates/bluetooth.x509.pem vendor/partner_modules/build/certificates/
+fi
